rename model files
