+++ mmocr/models/ner/classifer/__init__.py
@@ -1,3 +1,3 @@
-from .classifier import NerClassifier
+from .ner_classifier import NerClassifier
 
 __all__ = ['NerClassifier']
