+++ mmocr/models/ner/decoder/__init__.py
@@ -1,3 +1,3 @@
-from .decoder import FCDecoder
+from .fc_decoder import FCDecoder
 
 __all__ = ['FCDecoder']
