+++ mmocr/models/ner/encoder/__init__.py
@@ -1,3 +1,3 @@
-from .encoder import BertEncoder
+from .bert_encoder import BertEncoder
 
 __all__ = ['BertEncoder']
