+++ mmocr/models/ner/encoder/bert_encoder.py
