+++ mmocr/models/ner/loss/__init__.py
@@ -1,4 +1,4 @@
-from .label_smoothing import LabelSmoothingCrossEntropy
+from .label_smoothing_cross_entropy import LabelSmoothingCrossEntropy
 from .ner_loss import NerLoss
 
 __all__ = ['NerLoss', 'LabelSmoothingCrossEntropy']
