+++ mmocr/models/ner/loss/label_smoothing_cross_entropy.py
